memperbaiki search dan tampilan detail produk

+++ routes/web.php
@@ -3,8 +3,6 @@
 use App\Http\Controllers\DashboardProductController;
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\ProductController;
-use App\Models\Category;
-use App\Models\Product;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -35,7 +33,7 @@ Route::get('/dashboard/product', [DashboardProductController::class, 'index'])->
 Route::get('/dashboard/product/add', [DashboardProductController::class, 'create'])->middleware('admin');
 Route::post('/dashboard/product/add', [DashboardProductController::class, 'store'])->middleware('admin');
 // delete product
-Route::delete('/dashboard/product/{product:id}/delete', [DashboardProductController::class, 'destroy']);
+Route::delete('/dashboard/product/{product:id}/delete', [DashboardProductController::class, 'destroy'])->middleware('admin');
 
 // product for collection
 Route::get('/collection/{category:slug}', [ProductController::class, 'productCategory']);
